make separate texture example and separate rect and image to their own files
